Added "Are You Sure" prompt to all relevant functions

+++ functions/aks/analytics/uninstall.ps1
@@ -4,7 +4,10 @@ VerifyCurrentCluster $usage
 
 $analyticsName = ResourceGroupToAnalyticsName $selectedCluster.ResourceGroup
 
-Write-Info ("Uninstalling Log Analytics '$analyticsName' for current AKS cluster '$($selectedCluster.Name)'")
+if (AreYouSure)
+{
+    Write-Info ("Uninstalling Log Analytics '$analyticsName' for current AKS cluster '$($selectedCluster.Name)'")
 
-ExecuteCommand ("az aks disable-addons -a monitoring -g $($selectedCluster.ResourceGroup) -n $($selectedCluster.Name) $debugString")
-ExecuteCommand ("az resource delete --resource-type Microsoft.OperationalInsights/workspaces -g $($selectedCluster.ResourceGroup) -n $analyticsName $debugString")
+    ExecuteCommand ("az aks disable-addons -a monitoring -g $($selectedCluster.ResourceGroup) -n $($selectedCluster.Name) $debugString")
+    ExecuteCommand ("az resource delete --resource-type Microsoft.OperationalInsights/workspaces -g $($selectedCluster.ResourceGroup) -n $analyticsName $debugString")
+}
