+++ functions/aks/delete-pods.ps1
@@ -1,4 +1,4 @@
-# TODO: Add "Are You Sure" question.
+# Change to safely delete pods, by deleting 1 at a time and waiting for a new one to come up.
 param($deploymentName)
 
 $usage = Write-Usage "aks delete-pods [deployment]"
@@ -15,7 +15,9 @@ if (!$deployment)
 }
 
 $podNames = ExecuteQuery ("kubectl get po -l='app=$deploymentName' -o jsonpath='{.items[*].metadata.name}' $kubeDebugString")
-    
-Write-Info "Deleting pods in deployment '$deploymentName'"
-exit # Temp: Until TODO.
-ExecuteCommand ("kubectl delete pod $podNames $kubeDebugString")
+
+if (AreYouSure)
+{
+    Write-Info "Deleting pods in deployment '$deploymentName'"
+    ExecuteCommand ("kubectl delete pod $podNames $kubeDebugString")
+}
