+++ functions/aks/delete.ps1
@@ -1,8 +1,10 @@
-# TODO: Add "Are You Sure" question.
 $usage = Write-Usage "aks delete"
 
 VerifyCurrentCluster $usage
 
-Write-Info ("Deleting current AKS cluster")
-exit # Temp: Until TODO.
-ExecuteCommand ("az aks delete -n $($selectedCluster.Name) -g $($selectedCluster.ResourceGroup) $debugString")
+if (AreYouSure)
+{
+    Write-Info ("Deleting current AKS cluster")
+    
+    ExecuteCommand ("az aks delete -n $($selectedCluster.Name) -g $($selectedCluster.ResourceGroup) $debugString")
+}
