+++ functions/aks/monitoring/uninstall.ps1
@@ -2,7 +2,10 @@ $usage = Write-Usage "aks monitoring uninstall"
 
 VerifyCurrentCluster $usage
 
-Write-Info ("Uninstalling Monitoring (Prometheus & Grafana) from current AKS cluster '$($selectedCluster.Name)'")
+if (AreYouSure)
+{
+    Write-Info ("Uninstalling Monitoring (Prometheus & Grafana) from current AKS cluster '$($selectedCluster.Name)'")
 
-ExecuteCommand "helm delete grafana --purge $debugString"
-ExecuteCommand "helm delete prometheus --purge $debugString"
+    ExecuteCommand "helm delete grafana --purge $debugString"
+    ExecuteCommand "helm delete prometheus --purge $debugString"
+}
